Summary:

We skip the save_gpu_kernel if kernel is being saved already.
This would give us a more accurate Triton profiling result. The following trace shows before/after the change for a benchmarking of a trivial addmm:

Before:
 {F1889997034}

After:
 {F1889997398}

We can see that before the change, the benchmarking includes two parts,
(1) The overhead of our triton_heuristic call, which includes the save/get, and the (expensive) hash computation.
(2) The exact computation of Triton kernel.

We see that (1) accounts >50% of time, which makes kernel selection for profiling often choose aten kernels over Triton kernels.

Test Plan:
Existing OSS CI

IG_CTR (model id: 637300633) run:
```
TORCHINDUCTOR_FORCE_DISABLE_CACHES=1 TORCHINDUCTOR_MEMORY_PLANNING=1 TORCHINDUCTOR_UNIQUE_KERNEL_NAMES=1 TORCHINDUCTOR_MAX_AUTOTUNE=1 buck2 run mode/opt -c fbcode.platform010_cuda_version=12 -c fbcode.nvcc_arch=a100  caffe2/torch/fb/model_transform/experimental/benchmark:mts_gpu_benchmark -- --model-path=manifold://ads_storage_fblearner/tree/user/facebook/fblearner/predictor/637300633/0/gpu_lowering/input.predictor.disagg.gpu.merge --lower-backend="AOT_INDUCTOR"
```
Before:
```
== Benchmark Result for: Configuration(batch_iter=100, batch_size=2048, name='Test: AOTInductor', trt=False, ait=False, eager=False, jit=False, lower_dtype=torch.float16, accuracy_rtol=0.01, explicit_batch_dimension=True, report_aibench=False, verbose_profile=False, time_tensor_and_align=0, fx_time_tensor=False, use_cuda_graph=False, remove_passes=None, ait_profile=False, inductor=False, aot_inductor=True, aot_inductor_ep=False, num_threads=1, gpu_trace=True, op_level_profiling=False, additional_batch_size=[1, 512, 1024])
BS: 2048, MFLOPS/BS: 729.04, TFLOP/s: 68.53, Time per iter: 21.79ms, Threads: 1, QPS: 94000.50, Accuracy: True (rtol=0.01), AOT_INDUCTOR lowering duration: 545.33s
```
After:
```
== Benchmark Result for: Configuration(batch_iter=100, batch_size=2048, name='Test: AOTInductor', trt=False, ait=False, eager=False, jit=False, lower_dtype=torch.float16, accuracy_rtol=0.01, explicit_batch_dimension=True, report_aibench=False, verbose_profile=False, time_tensor_and_align=0, fx_time_tensor=False, use_cuda_graph=False, remove_passes=None, ait_profile=False, inductor=False, aot_inductor=True, aot_inductor_ep=False, num_threads=1, gpu_trace=True, op_level_profiling=False, additional_batch_size=[1, 512, 1024])
BS: 2048, MFLOPS/BS: 729.04, TFLOP/s: 72.40, Time per iter: 20.62ms, Threads: 1, QPS: 99303.27, Accuracy: True (rtol=0.01), AOT_INDUCTOR lowering duration: 528.94s
```

CMF (model id: 642218919) run:
```
TORCHINDUCTOR_FORCE_DISABLE_CACHES=1 TORCHINDUCTOR_MEMORY_PLANNING=1 TORCHINDUCTOR_UNIQUE_KERNEL_NAMES=1 TORCHINDUCTOR_MAX_AUTOTUNE=1 buck2 run mode/opt -c fbcode.platform010_cuda_version=12 -c fbcode.nvcc_arch=a100  caffe2/torch/fb/model_transform/experimental/benchmark:mts_gpu_benchmark -- --model-path=manifold://ads_storage_fblearner/tree/user/facebook/fblearner/predictor/642218919/58/gpu_lowering/input.predictor.disagg.gpu.merge --lower-backend="AOT_INDUCTOR"
```
Before:
```
== Benchmark Result for: Configuration(batch_iter=100, batch_size=2048, name='Test: AOTInductor', trt=False, ait=False, eager=False, jit=False, lower_dtype=torch.float16, accuracy_rtol=0.01, explicit_batch_dimension=True, report_aibench=False, verbose_profile=False, time_tensor_and_align=0, fx_time_tensor=False, use_cuda_graph=False, remove_passes=None, ait_profile=False, inductor=False, aot_inductor=True, aot_inductor_ep=False, num_threads=1, gpu_trace=True, op_level_profiling=False, additional_batch_size=[1, 512, 1024])
BS: 2048, MFLOPS/BS: 658.92, TFLOP/s: 82.71, Time per iter: 16.32ms, Threads: 1, QPS: 125515.74, Accuracy: True (rtol=0.01), AOT_INDUCTOR lowering duration: 1027.37s
```
After:
```
== Benchmark Result for: Configuration(batch_iter=100, batch_size=2048, name='Test: AOTInductor', trt=False, ait=False, eager=False, jit=False, lower_dtype=torch.float16, accuracy_rtol=0.01, explicit_batch_dimension=True, report_aibench=False, verbose_profile=False, time_tensor_and_align=0, fx_time_tensor=False, use_cuda_graph=False, remove_passes=None, ait_profile=False, inductor=False, aot_inductor=True, aot_inductor_ep=False, num_threads=1, gpu_trace=True, op_level_profiling=False, additional_batch_size=[1, 512, 1024])
BS: 2048, MFLOPS/BS: 658.92, TFLOP/s: 88.50, Time per iter: 15.25ms, Threads: 1, QPS: 134303.07, Accuracy: True (rtol=0.01), AOT_INDUCTOR lowering duration: 996.76s

```

Reviewed By: frank-wei

Differential Revision: D63441430